Update index.ts
The updateVehicle() function does not check if the vehicle already exists before updating it. This could allow non-existent vehicles to be updated.

The deleteVehicle() function does not check if the vehicle exists before deleting it. This could cause the function to fail if the vehicle does not exist.

The getVehicles() function does not handle the case where there are no vehicles in the storage. In this case, the function should return an empty Vec(Vehicle), not None.

The getVehicle() function does not handle the case where the vehicle does not exist. In this case, the function should return an Err({ NotFound: "..." }) result, not None.

The fuelStatus field in the VehicleData and Vehicle records is misspelled as fullStatus.

The updateVehicle() function does not update the updatedAt field for the vehicle.

The deleteVehicle() function does not return the deleted vehicle.

The updateVehicle() function could be improved by returning an Option<Vehicle> result instead of a Result(Vehicle, Error) result. This would allow the caller to distinguish between the cases where the vehicle was successfully updated and the cases where the vehicle did not exist or the update failed due to some other error.

The getVehicles() function could be improved by returning a Vec(Vehicle) result even if there are no vehicles in the storage. This would make the function more consistent with the other functions in the canister, which all return Ok({}) results if there is no data to return.

The getVehicle() function could be improved by returning an Err({ NotFound: "..." }) result if the vehicle does not exist. This would make the function more consistent with the other functions in the canister, which all return Err({ ... }) results if there is an error.